Moving the LocationRepository to the infra package
