+++ src/Location/LocationRepositoryInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CartBooking\Location;
+
+interface LocationRepositoryInterface
+{
+    /**
+     * @return \Generator|Location[]
+     */
+    public function findAll();
+
+    public function findById(int $id);
+}
