+++ src/Provider/RepositoryProvider.php
@@ -4,13 +4,10 @@ namespace CartBooking\Provider;
 
 use Bigcommerce\Injector\InjectorServiceProvider;
 use CartBooking\Booking\BookingRepository;
-use CartBooking\Location\LocationRepository;
+use CartBooking\Infrastructure\Persistence\Doctrine\Repository\DoctrineLocationRepository;
 use CartBooking\Publisher\PublisherRepository;
 use CartBooking\Shift\ShiftRepository;
-use Doctrine\DBAL\Connection;
-use Doctrine\ORM\EntityManager;
 use Pimple\Container;
-use Silex\Provider\DoctrineServiceProvider;
 
 class RepositoryProvider extends InjectorServiceProvider
 {
@@ -51,7 +48,7 @@ class RepositoryProvider extends InjectorServiceProvider
         });
         $this->alias('repository.shift', ShiftRepository::class);
 
-        $this->autoBind(LocationRepository::class);
-        $this->alias('repository.location', LocationRepository::class);
+        $this->autoBind(DoctrineLocationRepository::class);
+        $this->alias('repository.location', DoctrineLocationRepository::class);
     }
 }
